+++ lib/type_def/type_def.dart
@@ -1,0 +1,3 @@
+typedef FieldValidator<T> = String? Function(T? value);
+
+typedef TextFieldValidator = String? Function(String? value);