+++ lib/type_defs/type_defs.dart
@@ -1,1 +1,0 @@
-typedef FieldValidator<T> = String? Function(T? value);