+++ lib/view_models/field_cubits/text_field_cubit/text_field_cubit.dart
@@ -2,6 +2,7 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:reusable_editor/type_def/type_def.dart';
 
 part 'text_field_state.dart';
 
@@ -22,34 +23,39 @@ class TextFieldCubit extends Cubit<TextFieldState> {
   }
 
   void _handleTextChanged() {
-    // Keep Cubit state updated when user types
     final value = controller.text;
     if (value != state.value) {
-      final error = state.validator?.call(value);
-      emit(state.copyWith(value: value, errorText: error));
+      final error = state.isDirty ? state.validator?.call(value) : null;
+      emit(state.copyWith(value: value, errorText: error, isDirty: true));
     }
   }
 
   void onChanged(String? value) {
     final newValue = value?.trim() ?? '';
-    // Update Cubit + TextController together
     if (newValue != state.value) {
       controller.text = newValue;
       controller.selection = TextSelection.collapsed(offset: newValue.length);
-      final error = state.validator?.call(newValue);
-      emit(state.copyWith(value: newValue, errorText: error));
+      final error = state.isDirty ? state.validator?.call(newValue) : null;
+      emit(state.copyWith(value: newValue, errorText: error, isDirty: true));
     }
   }
 
-  String? validate() => _validateAndEmit(state.value);
-
-  FormFieldValidator<String> get formFieldValidator => _validateAndEmit;
-
-  String? _validateAndEmit(String? value) {
-    final error = state.validator?.call(value);
-    emit(state.copyWith(errorText: error));
+  String? validate({bool force = false}) {
+    // If force is true (create), validate all
+    // If force is false (update), validate only dirty fields
+    final shouldValidate = force || state.isDirty;
+    final error = shouldValidate ? state.validator?.call(state.value) : null;
+    emit(
+      state.copyWith(
+        errorText: error,
+        isDirty: shouldValidate || state.isDirty,
+      ),
+    );
     return error;
   }
+
+  FormFieldValidator<String> get formFieldValidator =>
+      (value) => validate();
 
   void reset({bool validate = false}) {
     controller.text = _initialValue ?? '';
@@ -58,12 +64,14 @@ class TextFieldCubit extends Cubit<TextFieldState> {
     );
 
     final error = validate ? state.validator?.call(_initialValue) : null;
-    emit(state.copyWith(value: _initialValue, errorText: error));
+    emit(
+      state.copyWith(value: _initialValue, errorText: error, isDirty: false),
+    );
   }
 
   void clear() {
     controller.clear();
-    emit(state.copyWith(value: null, errorText: null));
+    emit(state.copyWith(value: null, errorText: null, isDirty: false));
   }
 
   void dispose() {