+++ lib/view_models/generic_cubits/field_cubit/field_state.dart
@@ -3,9 +3,15 @@ part of 'field_cubit.dart';
 class FieldState<T> extends Equatable {
   final T? value;
   final String? errorText;
-  final FieldValidator<T>? validator; // ✅ updated type
+  final FieldValidator<T>? validator;
+  final bool isDirty;
 
-  const FieldState({this.value, this.errorText, this.validator});
+  const FieldState({
+    this.value,
+    this.errorText,
+    this.validator,
+    this.isDirty = false,
+  });
 
   bool get isValid => errorText == null;
 
@@ -13,21 +19,23 @@ class FieldState<T> extends Equatable {
     T? value,
     String? errorText,
     FieldValidator<T>? validator,
-    bool clearError = false,
+    bool? isDirty,
   }) {
     return FieldState<T>(
       value: value ?? this.value,
-      errorText: clearError ? null : errorText ?? this.errorText,
+      errorText: errorText,
       validator: validator ?? this.validator,
+      isDirty: isDirty ?? this.isDirty,
     );
   }
 
   FieldState<T> init() => FieldState<T>(
     value: null,
     validator: validator,
-    errorText: validator?.call(null),
+    errorText: null,
+    isDirty: false,
   );
 
   @override
-  List<Object?> get props => [value, errorText, validator];
+  List<Object?> get props => [value, errorText, validator, isDirty];
 }